+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `String.lcfirst()`
 - Added `String.lower()`
 - Added `String.plural()`
+- Added `String.singular()`
 - Added `String.take()`
 - Added `String.ucfirst()`
 - Added `String.upper()`